requirements.txt
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
cssselect==1.3.0
dj-rest-auth==7.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-spectacular==0.28.0
fake-useragent==2.2.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
lxml_html_clean==0.4.2
mongoengine==0.29.1
mysqlclient==2.2.7
packaging==25.0
parse==1.20.2
psycopg2-binary==2.9.10
pyee==11.1.1
PyJWT==2.10.1
pymongo==4.14.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-html==0.10.0
rpds-py==0.27.1
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
zipp==3.23.0
